checkRunner-sender-mail: changed subscribing of event (send receipt to email)
