+++ src/main/java/by/markov/services/event/Observer.java
@@ -1,10 +1,8 @@
 package by.markov.services.event;
 
-import by.markov.models.Products;
-
 import javax.mail.MessagingException;
 import java.io.IOException;
 
 public interface Observer {
-    void handleEvent(Products product, double sum) throws MessagingException, IOException;
+    void handleEvent(String message) throws MessagingException, IOException;
 }
