+++ src/main/java/by/markov/services/receipts/ReceiptsHandler.java
@@ -7,7 +7,7 @@ import by.markov.models.Receipt;
 import by.markov.models.Supermarket;
 import by.markov.services.cashiers.CashierCreator;
 import by.markov.services.exceptions.CashierHandlerException;
-import by.markov.services.mailsender.Sender;
+import by.markov.services.mailsender.SenderImplObserver;
 import by.markov.services.supermarkets.SupermarketCreator;
 
 import javax.mail.MessagingException;
@@ -33,7 +33,6 @@ public class ReceiptsHandler {
         SupermarketCreator supermarketCreator = new SupermarketCreator();
         Supermarket supermarket = supermarketCreator.createDefaultSupermarket();
         CashierCreator cashierCreator = new CashierCreator();
-        StringBuilder messageToMail = new StringBuilder();
         try {
             Cashier cashier = cashierCreator.createDefaultCashier(CashierCreator.DEFAULT_CASHIER_NAME);
             ReceiptPrinter receiptPrinter = new ReceiptPrinter();
@@ -44,16 +43,9 @@ public class ReceiptsHandler {
                     .append(receiptPrinter.printShopBasket(receipt.getShopBasket())).append('\n')
                     .append(receiptPrinter.printTotalSum(receipt));
             fileWriter.close();
-            messageToMail.append(printReceiptTitle()).append('\n')
-                    .append(receiptPrinter.printSupermarket(supermarket)).append('\n')
-                    .append(receiptPrinter.printCashierWithInfo(cashier)).append('\n')
-                    .append(receiptPrinter.printShopBasket(receipt.getShopBasket())).append('\n')
-                    .append(receiptPrinter.printTotalSum(receipt));
         } catch (CashierHandlerException | IOException e) {
             System.out.println(e.getMessage());
         }
-        Sender sender = new Sender();
-        sender.sendCheckToMail(messageToMail.toString());
         return file;
     }
 
@@ -76,5 +68,9 @@ public class ReceiptsHandler {
         if (!ReceiptRunner.isAppRunViaCommandLine) {
             handler.display(file);
         }
+
+        SenderImplObserver senderImplObserver = new SenderImplObserver();
+        receipt.addObserver(senderImplObserver);
+        receipt.notifyObservers();
     }
 }
